Added bot for flax. Target quest objects are now highlighted in the world with circles Dryer works correctly with input barter Updated the formula for calculating the rating of pigs The butcher can now correctly set the drop threshold for entrails and intestines

+++ src/nurgling/bots/settings/ThresholdSetter.java
@@ -16,7 +16,7 @@ public class ThresholdSetter extends Widget {
         add(new Button(50,"Set"){
             @Override
             public void click() {
-                AreasID.set(id,Integer.valueOf(entry.text()));
+                AreasID.set(id,Double.valueOf(entry.text()));
                 AreasID.write();
             }
         },prev.pos("ur").add(5,-2));
